spring-course: implementando paciente DTO, Repository, Controller e migration

+++ spring-boot/spring-course/src/main/java/api/medico/MedicoRepository.java
@@ -2,4 +2,4 @@ package api.medico;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MedicoRepository extends JpaRepository<Medico, Long> {}
+public interface MedicoRepository extends JpaRepository<Medico, Integer> {}
